Don't expect TCP_INFO and clock_gettime() to be portable
Sadly, TCP_INFO is only available on Linux and FreeBSD. Other systems make
it easy to retrieve interface-specific parameters, but not the parameters
for a given socket, except by inspecting PF states.

clock_gettime() is available on all BSD systems, but not on OSX.

mach_absolute_time() can be used there, but since it was only used to
display the TCP socket info, and TCP_INFO doesn't exist on OSX, let's
just ignore that altogether for now.

That's pretty sad but at least, it makes glorytun usable on !__linux__